really fix for the Calender
git-svn-id: https://svn.apache.org/repos/asf/incubator/wicket/branches/wicket-1.x@507220 13f79535-47bb-0310-9956-ffa450edef68

+++ wicket/src/main/java/wicket/util/io/ClassStreamHandler.java
@@ -293,7 +293,7 @@ public final class ClassStreamHandler
 				}
 				else if (value instanceof Long)
 				{
-					woos.getOutputStream().writeLong(((Long)value).intValue());
+					woos.getOutputStream().writeLong(((Long)value).longValue());
 					continue;
 				}
 				else if (value instanceof Float)
